Finished member unban log, #30

+++ src/modules/Mod/ban.ts
@@ -1,8 +1,9 @@
-import { Formatters, Guild, Permissions } from "discord.js"
+import { Formatters, Guild, GuildMember, Permissions, User } from "discord.js"
 import { Snowflake } from "discord-api-types"
 
 import { GuildBanModel } from "../../database/schemas/GuildBans"
 import Logs from "../Logs"
+import { Language } from "types"
 
 export const Ban = {
     add: async (targetId: Snowflake, guildId: Snowflake, executorId: Snowflake, reason?: string, time?: number) => {
@@ -30,7 +31,7 @@ export const Ban = {
 
         return { result }
     },
-    remove: async (targetId: Snowflake, guildId: Snowflake, by: Snowflake, reason?: string) => {
+    remove: async (targetId: Snowflake, guildId: Snowflake, executorId: Snowflake, reason?: string) => {
         if(!targetId) return { error: "Missing targetId" }
         if(!guildId) return { error: "Missing guildId" }
         const guild = await client.guilds.fetch(guildId).catch(e => {});
@@ -38,9 +39,11 @@ export const Ban = {
         if(!guild.me?.permissions.has(Permissions.FLAGS.BAN_MEMBERS)) return { error: "missingPermission", perms: new Permissions([Permissions.FLAGS.BAN_MEMBERS]).toArray() }
         const target = await guild.members.fetch(targetId).catch(() => {}) || await client.users.fetch(targetId).catch(() => {})
         if(!target) return { error: "Wrong targetId" }
+        const language = guild.preferredLocale === 'pl' ? 'pl' : 'en'
         
         const result = await guild.bans.remove(targetId, reason).catch(e => {})
         if(!result) return { error: "notBanned" }
+        await Logs.log('members', 'unban', guild.id, { member: target, customs: { moderatorMention: `<@${executorId}>`, moderatorId: executorId, reason: reason || t('general.none', language) } })
 
         return { result };
     }
@@ -61,8 +64,11 @@ async function saveBan(userId: Snowflake, guild: Guild, executorId: Snowflake, r
     await Logs.log('members', 'ban', guild.id, { member: user, customs: { moderatorMention: `<@${executorId}>`, moderatorId: executorId,  reason: reason || t('general.none', language), unbanDate: timestamp ? Formatters.time(Math.floor(timestamp / 1000)) : t('general.never', language), unbanDateR: timestamp ? Formatters.time(Math.floor(timestamp / 1000), 'R') : " " }})
 
     setTimeout(async () => {
-        await guild.bans.remove(userId, `Timed ban ended, reason: ${reason || 'none'}`).catch(e => {})
+        if(!guild.me?.permissions.has(Permissions.FLAGS.BAN_MEMBERS)) return
+        const unbanReason = t('autoMod.ban.removeReason', language, { reason: reason || t('general.none', language) })
+        await guild.bans.remove(userId, unbanReason).catch(e => {})
         await GuildBanModel.findOneAndDelete({ userId, guildId: guild.id });
+        await Logs.log('members', 'unban', guild.id, { member: user, customs: { moderatorMention: `<@${client.user?.id}>`, moderatorId: client.user?.id, reason: unbanReason }})
     }, time);
 
     return true;
